Improved table output for complex resource properties.
HMC resource properties with complex types (arrays or nested objects)
so far were generated as the `repr()` string of the data value.

With this change, they are now generated as nested tables, wherei
possible. Specifically, the following constructs are used for these
inner tables, dependent on the output format:

* table/psql, simple, plain: Nested table in plain format.
* rst: Nested table in RST full table format (tabulate
  'grid' format).
* mediawiki: Nested table in mediawiki format.
* html: Nested table in html format.
* latex: Still the `repr()` string, as without this change. While
  LaTex very well supports nested tables, using them would require
  larger changes in the "tabulate" package to turn off the extra
  LaTex escaping it performs.

This change depends on the recently released version 0.8.1 of the
Python "tabulate" package. Adjusted the dependencies accordingly.

Signed-off-by: Andreas Maier <andreas.r.maier@gmx.de>