Add support for `clear_indicator` and `store_status_indicator` in LPAR Load
Add support for `clear_indicator` and `store_status_indicator` in the
LPAR Load operation. As the default value for `clear-indicator` in the
zhmcclient libary is `True` we must add a separate flag so you can
disable it (see
https://click.palletsprojects.com/en/7.x/options/#boolean-flags for
details).

Dependency:

This change depends on support for `clear_indicator` and
`store_status_indicator` parameters in `Lpar.load()`.

Signed-off-by: Marc Hartmayer <mhartmay@linux.ibm.com>